requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argcomplete==2.1.1
atila==0.26.8
backports.zoneinfo==0.2.1
boto3==1.26.93
botocore==1.29.93
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
dacite==1.7.0
Deprecated==1.2.13
dill==0.3.6
dnn==0.7.4
dnspython==2.3.0
easydict==1.10
eventlet==0.33.3
exif==1.6.0
fdet==0.2.1
fiftyone==0.19.1
fiftyone-brain==0.10.0
fiftyone-db==0.4.0
filelock==3.9.0
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.39.2
future==0.18.3
glob2==0.7
google-auth==2.16.2
google-auth-oauthlib==0.4.6
graphql-core==3.2.3
greenlet==2.0.2
grpcio==1.51.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.1
humanfriendly==10.0
hypercorn==0.14.3
hyperframe==6.0.1
hyperopt==0.2.7
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
insightface==0.7.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kaleido==0.2.1
kiwisolver==1.4.4
lazy_loader==0.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mongoengine==0.24.2
motor==3.1.1
mpmath==1.3.0
ndjson==0.3.1
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.5.3
patool==1.12
Pillow==9.4.0
plotly==5.13.1
plum-py==0.8.5
pprintpp==0.4.0
prettytable==3.6.0
priority==2.0.0
protobuf==3.20.3
psutil==5.9.4
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3986==1.5.0
rs4==0.3.21.4
rsa==4.9
s3transfer==0.6.0
sahi==0.11.12
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.17.0
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
skitai==0.56.8
sniffio==1.3.0
sortedcontainers==2.4.0
sqlphile==0.9.7
sse-starlette==0.10.3
sseclient-py==1.7.2
starlette==0.20.4
strawberry-graphql==0.138.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminaltables==3.1.10
tfserver==0.4.9
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.15
timm==0.6.12
tokenizers==0.13.2
toml==0.10.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
ttictoc==0.5.6
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ultralytics==8.0.43
ultralyticsplus==0.0.28
universal-analytics-python3==1.1.1
urllib3==1.26.14
voxel51-eta==0.8.4
wcwidth==0.2.6
webrtcvad==2.0.10
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.15.0

ipython~=8.11.0
pycocotools~=2.0.6